agentic-shared(move-session-url,mechanical_provable): move resolve_session_url into miles.rollout.agentic.session
The session-URL shaping (append /v1, rewrite the host to
MILES_ROUTER_EXTERNAL_HOST) leaves openenv_agent_function.py unchanged; the
OpenEnv leg imports it. The NeMo Gym and Harbor legs switch to it next.
