+++ miles/rollout/agentic/session.py
@@ -1,0 +1,15 @@
+"""The session-scoped policy URL an agent function hands to its agent."""
+
+import os
+from urllib.parse import urlparse, urlunparse
+
+
+def resolve_session_url(base_url: str) -> str:
+    """Build the OpenAI-compatible policy URL, rewriting host for off-cluster agents."""
+    session_url = f"{base_url}/v1"
+    external_host = os.getenv("MILES_ROUTER_EXTERNAL_HOST")
+    if external_host:
+        parsed = urlparse(session_url)
+        netloc = f"{external_host}:{parsed.port}" if parsed.port else external_host
+        session_url = urlunparse(parsed._replace(netloc=netloc))
+    return session_url
